### 2022-10-13 #4850 @editorialbot
[PRE REVIEW]: nnTensor: An R package for non-negative matrix/tensor decomposition
**Submitting author:** @kokitsuyuzaki (<a href="http://orcid.org/0000-0003-3797-2148">Koki Tsuyuzaki</a>)
**Repository:** https://github.com/rikenbit/nnTensor
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.9
