### 2022-10-13 #4850 @kthyng
Hi @kokitsuyuzaki and thanks for your submission. I'm not very familiar with R so maybe I am missing it, but where are the docs for your package?
