### 2022-10-14 #4850 @kokitsuyuzaki
@kthyng
Thank you for facilitating the editing of this manuscript.

> where are the docs for your package?

.Rd files in man/ directory are the documents of R package
https://github.com/rikenbit/nnTensor/tree/master/man

and user can check the usage, argument, and output of all the functions in this page.
https://cran.rstudio.com/web/packages/nnTensor/nnTensor.pdf
