### 2022-10-14 #4850 @kokitsuyuzaki
For the potential reviewers, I recommend `oneilsh`, who is the author of [tidyTensor](https://joss.theoj.org/papers/10.21105/joss.03543).
In the list, I think, `jeffhammond`,` jamiehadd`, `yhtang`, and `meandmytram` are suitable for the review of tensor decomposition.
