### 2022-10-23 #4850 @kokitsuyuzaki
@kthyng Excuse me for being unfamiliar with this JOSS submission system, but now I just have to wait until some reviewer is assigned or is there anything I have to do?
