+++ R/NTF.R
@@ -295,7 +295,7 @@ NTF <- function(X, M=NULL, pseudocount=.Machine$double.eps,
                     A[[n]] <- A[[n]][orderA, ]
                 }
             } else if (init == "ALS") {
-                Xn <- cs_unfold(X, m = 1)@data
+                Xn <- cs_unfold(X, m = n)@data
                 res.svd <- svd(Xn)
                 A[[n]] <- t(.positive(res.svd$v[, seq(rank)]))
                 orderA <- order(sapply(res.svd$d[seq(rank)], function(x) {
