+++ tests/testthat/test_NTF.R
@@ -4,8 +4,9 @@
 X <- toyModel("CP")
 
 out1_1 <- NTF(X, rank=3, algorithm="Frobenius", num.iter=2)
-out1_2 <- NTF(X, rank=3, algorithm="Frobenius", init="ALS", num.iter=2)
-out1_3 <- NTF(X, rank=3, algorithm="Frobenius", init="Random", num.iter=2)
+out1_2 <- NTF(X, rank=3, algorithm="Frobenius", init="ABS-SVD", num.iter=2)
+out1_3 <- NTF(X, rank=3, algorithm="Frobenius", init="ALS", num.iter=2)
+out1_4 <- NTF(X, rank=3, algorithm="Frobenius", init="Random", num.iter=2)
 out2 <- NTF(X, rank=3, algorithm="KL", num.iter=2)
 out3 <- NTF(X, rank=3, algorithm="IS", num.iter=2)
 out4 <- NTF(X, rank=3, algorithm="Pearson", num.iter=2)
@@ -20,6 +21,7 @@ out11 <- NTF(X, rank=3, algorithm="Beta-HALS", num.iter=2)
 expect_equivalent(length(out1_1), 6)
 expect_equivalent(length(out1_2), 6)
 expect_equivalent(length(out1_3), 6)
+expect_equivalent(length(out1_4), 6)
 expect_equivalent(length(out2), 6)
 expect_equivalent(length(out3), 6)
 expect_equivalent(length(out4), 6)
