### 2022-11-23 #4850 @Stat-Cook
@prashjha I'd be happy to help out.  Not a specialist in tensor decomposition but I've used non-negative matrix decomp in anger in a few places.  Could I get some quick guidance as to the process?  I checked one of the other issues and there's a checklist?
