### 2022-12-04 #4850 @kokitsuyuzaki
@Stat-Cook, thank you for accepting the review of my paper.

@prashijha, thank you for finding a peer reviewer. Do you have the guidance or the checklist for @Stat-Cook?
