### 2022-12-15 #5015 @editorialbot
[REVIEW]: nnTensor: An R package for non-negative matrix/tensor decomposition
**Submitting author:** @kokitsuyuzaki (<a href="http://orcid.org/0000-0003-3797-2148">Koki Tsuyuzaki</a>)
**Repository:** https://github.com/rikenbit/nnTensor
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.14
