### 2022-12-15 rikenbit/nnTensor#1 @Stat-Cook
Documented action of plotTensor3D
[Lovely package by the way - please excuse if these comments read as highly kurt I'm trying to be concise]

The use of plotTensor3D given in the JOSS paper does not reproduce on my system - instead get a semi blank image.  May be other dependencies needed.

![image](https://user-images.githubusercontent.com/51404549/207908982-d4036e9a-96ab-4591-9204-6d7992af272c.png)
