### 2022-12-15 #5015 @Stat-Cook
@kokitsuyuzaki can I confirm how you intend the paper to be viewed?  When I visit paper.md there are several unprocessed tags, either an error in the github markdown or I'm not viewing as intended.
