### 2022-12-15 #5015 @Stat-Cook
@kokitsuyuzaki you have both a license file and a license in the read me.  Is this intentional, and if not please reduce for clarity.
