### 2022-12-16 #5015 @kokitsuyuzaki
> @kokitsuyuzaki can I confirm how you intend the paper to be viewed? When I visit paper.md there are several unprocessed tags, either an error in the github markdown or I'm not viewing as intended.

It might be because of the difference between the markdown of JOSS and that of GitHub.
To review my manuscript, please check the latest PDF generated by `@editorialbot generate pdf`.
cf. https://github.com/openjournals/joss-reviews/issues/5015#issuecomment-1353379430
