### 2022-12-16 #5015 @kokitsuyuzaki
> @kokitsuyuzaki you have both a license file and a license in the read me. Is this intentional, and if not please reduce for clarity.

@Stat-Cook Sorry, I'll remove the licence section in README.md later.
