### 2022-12-16 rikenbit/nnTensor#1 @kokitsuyuzaki
Could you send the message generated by `sessionInfo()` in R console?
