### 2022-12-23 rikenbit/nnTensor#1 @kokitsuyuzaki
I couldn't reproduce your situation in my environment.
I checked the function on Mac desktop and Linux (CentOS) server but the plot is properly generated.

The function is tested here,
https://github.com/rikenbit/nnTensor/blob/db7feffb1e8618ad327d210ffda4a3a6b41957b0/tests/testthat.R#L16
and at least the figure is properly generated on 13 machines.
https://cran.rstudio.com/web/checks/check_results_nnTensor.html
