### 2022-12-23 rikenbit/nnTensor#1 @kokitsuyuzaki
Perhaps, is it related to XQuartz (https://www.xquartz.org)?

If you have not already, could you install XQuartz on your machine and draw the figure again?
https://cran.r-project.org/web/packages/nat/vignettes/Installation.html
