### 2022-12-23 #5015 @kokitsuyuzaki
@Stat-Cook l've removed the license section in README.md.
https://github.com/rikenbit/nnTensor
