### 2022-12-23 #5015 @kokitsuyuzaki
@Stat-Cook In my environment, `plotTensor3D` works properly. Could you give me some hint to reproduce the error (e.g., `sessionInfo()`)?
https://github.com/rikenbit/nnTensor/issues/1
