+++ paper/paper.bib
@@ -3,7 +3,8 @@
     Author = {{Kolda}, T. G. and {Bader}, B. W.},
     Journal = {SIAM Review},
     Title = {{Tensor Decompositions and Applications}},
-    Year = 2009
+    Year = 2009,
+    doi = {10.1137/07070111X}
 }
 
 @article{tensorlab,
@@ -11,7 +12,8 @@
     Author = {{Vervliet}, N. and {Debals}, O. and {De Lathauwer}, L.},
     Journal = {2016 50th Asilomar Conference on Signals, Systems and Computers},
     Title = {{Tensorlab 3.0 - Numerical optimization strategies for large-scale constrained and coupled matrix/tensor factorization}},
-    Year = 2016
+    Year = 2016,
+    doi = {10.1109/acssc.2016.7869679}
 }
 
 @article{tensorly,
@@ -27,13 +29,14 @@
     Author = {{Li}, J. and {Bien}, J. and {Wells}, M. T.},
     Journal = {Journal of Statistical Software},
     Title = {{rTensor: An R Package for Multidimensional Array (Tensor) Unfolding, Multiplication, and Decomposition}},
-    Year = 2018
+    Year = 2018,
+    doi = {10.18637/jss.v087.i10}
 }
 
 @book{nonnegative,
     url = {https://onlinelibrary.wiley.com/doi/book/10.1002/9780470747278},
     Author = {{Cichocki}, A. and {Zdunek}, F. and {Phan}, A. H. and {Amari}, S.},
-    Booktitle = {Nonnegative Matrix and Tensor Factorizations: Applications to Exploratory Multi-way Data Analysis and Blind Source Separation, by Andzej Cichocki, Fafal Zdunek, Anh Huy Phan, and Shun-ichi Amari.~ISBN 978-0470746660.~Published by Wiley, 2009.},
+    Booktitle = {Nonnegative Matrix and Tensor Factorizations: Applications to Exploratory Multi-way Data Analysis and Blind Source Separation},
     Publisher = {Wiley},
     Year = 2009
 }
@@ -43,7 +46,8 @@
     Author = {{Čopar}, A. and {Zupan}, B. and {Zitnik}, M.},
     Journal = {PLOS ONE},
     Title = {{Fast Optimization of Non-Negative Matrix Tri-Factorization}},
-    Year = 2019
+    Year = 2019,
+    doi = {10.1371/journal.pone.0217994}
 }
 
 @article{sinmf,
@@ -51,7 +55,8 @@
     Author = {{Badea}, L.},
     Journal = {Pacific Symposium on Biocomputing 2008},
     Title = {{Extracting Gene Expression Profiles Common to Colon and Pancreatic Adenocarcinoma using Simultaneous nonnegative matrix factorization}},
-    Year = 2009
+    Year = 2009,
+    doi = {10.1142/9789812776136_0027}
 }
 
 @article{jnmf,
@@ -59,5 +64,6 @@
     Author = {{Yang}, Z. and {Michailidis}, G.},
     Journal = {Bioinformatics},
     Title = {{A non-negative matrix factorization method for detecting modules in heterogeneous omics multi-modal data}},
-    Year = 2016
+    Year = 2016,
+    doi = {10.1093/bioinformatics/btv544}
 }
