### 2022-12-23 #5015 @kokitsuyuzaki
As whedon mentioned it, I added DOIs for the references as possible.
https://github.com/rikenbit/nnTensor/blob/master/paper/paper.bib
