### 2022-12-31 #5015 @oneilsh
@kokitsuyuzaki I've started taking a look at nnTensor; impressive work! My initial comments are around some of the documentation and example usage. The help for the functions (I'm looking at the doc for `NTD`, others look similar) seems to assume that the `num_modes` of the input tensor will be 3 (with explicit references to I1, I2, I3, J1, J2, and J3), but I see that the methods are implemented generally and work with tensors with more modes. I think the help docs should reflect the general case, especially the argument definitions and return value definitions. You could then add a Details section that provides more in-depth explanations on specific tensor shapes if it helps. The example usage is also quite sparse, and would benefit from examples on different input and core tensor shapes and using/interpreting the output.
