+++ tests/testthat/test_mask_NMF.R
@@ -6,20 +6,20 @@ M2[,] <- rbinom(length(M2), 1, 0.5)
 
 
 
-out1 <- NMF(X, M=M1, J=3, algorithm="Frobenius", num.iter=2)
-out2 <- NMF(X, M=M1, J=3, algorithm="KL", num.iter=2)
-out3 <- NMF(X, M=M1, J=3, algorithm="IS", num.iter=2)
-out4 <- NMF(X, M=M1, J=3, algorithm="Pearson", num.iter=2)
-out5 <- NMF(X, M=M1, J=3, algorithm="Hellinger", num.iter=2)
-out6 <- NMF(X, M=M1, J=3, algorithm="Neyman", num.iter=2)
-out7 <- NMF(X, M=M1, J=3, algorithm="Alpha", num.iter=2)
-out8 <- NMF(X, M=M1, J=3, algorithm="Beta", num.iter=2)
-out9 <- NMF(X, M=M1, J=3, algorithm="PGD", num.iter=2)
-out10 <- NMF(X, M=M1, J=3, algorithm="Projected", num.iter=2)
-out11 <- NMF(X, M=M1, J=3, algorithm="NHR", num.iter=2)
-out12 <- NMF(X, M=M1, J=3, algorithm="DTPP", num.iter=2)
-out13 <- NMF(X, M=M1, J=3, algorithm="Orthogonal", num.iter=2)
-out14 <- NMF(X, M=M1, J=3, algorithm="OrthReg", num.iter=2)
+out1 <- NMF(X, M=M1, J=3, algorithm="Frobenius", num.iter=10)
+out2 <- NMF(X, M=M1, J=3, algorithm="KL", num.iter=10)
+out3 <- NMF(X, M=M1, J=3, algorithm="IS", num.iter=10)
+out4 <- NMF(X, M=M1, J=3, algorithm="Pearson", num.iter=10)
+out5 <- NMF(X, M=M1, J=3, algorithm="Hellinger", num.iter=10)
+out6 <- NMF(X, M=M1, J=3, algorithm="Neyman", num.iter=10)
+out7 <- NMF(X, M=M1, J=3, algorithm="Alpha", num.iter=10)
+out8 <- NMF(X, M=M1, J=3, algorithm="Beta", num.iter=10)
+out9 <- NMF(X, M=M1, J=3, algorithm="PGD", num.iter=10)
+out10 <- NMF(X, M=M1, J=3, algorithm="Projected", num.iter=10)
+out11 <- NMF(X, M=M1, J=3, algorithm="NHR", num.iter=10)
+out12 <- NMF(X, M=M1, J=3, algorithm="DTPP", num.iter=10)
+out13 <- NMF(X, M=M1, J=3, algorithm="Orthogonal", num.iter=10)
+out14 <- NMF(X, M=M1, J=3, algorithm="OrthReg", num.iter=10)
 
 expect_equivalent(length(out1), 10)
 expect_equivalent(length(out2), 10)
@@ -38,20 +38,20 @@ expect_equivalent(length(out14), 10)
 
 
 
-out15 <- NMF(X, M=M2, J=3, algorithm="Frobenius", num.iter=2)
-out16 <- NMF(X, M=M2, J=3, algorithm="KL", num.iter=2)
-out17 <- NMF(X, M=M2, J=3, algorithm="IS", num.iter=2)
-out18 <- NMF(X, M=M2, J=3, algorithm="Pearson", num.iter=2)
-out19 <- NMF(X, M=M2, J=3, algorithm="Hellinger", num.iter=2)
-out20 <- NMF(X, M=M2, J=3, algorithm="Neyman", num.iter=2)
[…]
