### 2023-02-15 rikenbit/nnTensor#1 @kokitsuyuzaki
I found that this error occurs when plotting with R-Studio.

![スクリーンショット 2023-02-15 16 00 02](https://user-images.githubusercontent.com/2173745/219047442-55dd09ee-0fdc-4c99-b471-2f2e7094bfc6.png)

The cause was that when plotting with plot3D::scatter3D, the area where 0 is specified in col becomes transparent in R.app or R in CLI but in R-Studio, it seems to be drawn as white for some reason.
Now I was able to reproduce the Figure in the JOSS paper in R-Studio by explicitly specifying a color with transparency.
In my environment, this bug was resolved in nnTensor (v-1.1.12), so please try in that environment.
If the bug cannot resolved, please open the Issue again.
