### 2023-02-15 #5015 @kokitsuyuzaki
@Stat-Cook Sorry for the late reply. A bug regarding `plotTensor3D` has been resolved. For more information, please see the comment in the following Issue.
https://github.com/rikenbit/nnTensor/issues/1
