### 2023-02-15 #5015 @kokitsuyuzaki
@oneilsh Sorry for the late reply.

> I think the help docs should reflect the general case, especially the argument definitions and return value definitions.

Thank you for your suggestion.
I have updated the manuals to explain that any order tensor can be specified as the input.
https://github.com/rikenbit/nnTensor/blob/master/man/NTD.Rd#L30
https://github.com/rikenbit/nnTensor/blob/master/man/NTF.Rd#L25
https://github.com/rikenbit/nnTensor/blob/master/man/recTensor.Rd#L20

> You could then add a Details section that provides more in-depth explanations on specific tensor shapes if it helps. The example usage is also quite sparse, and would benefit from examples on different input and core tensor shapes and using/interpreting the output.

I have created three vignettes to explain the way to use `NMF`, `NMTF`, `siNMF`, `jNMF`, `NTF`, and `NTD`.
https://github.com/rikenbit/nnTensor/tree/master/vignettes
https://cran.r-project.org/web/packages/nnTensor/vignettes/nnTensor-1.html
https://cran.r-project.org/web/packages/nnTensor/vignettes/nnTensor-2.html
https://cran.r-project.org/web/packages/nnTensor/vignettes/nnTensor-3.html

These vignettes are designed for a single matrix, multiple matrices, and a single tensor, respectively, and the third vignette explains that `NTD` can also work as NTD-2/NTD-1 which extract factor matrices only from specific modes. In the third vignette, I also explain that `NTF` and `NTD` can be applied to higher-order tensors.
