### 2023-03-08 #5015 @oneilsh
@kokitsuyuzaki the vignettes look great and even provide multiple examples. The only remaining issue I see is the lack of community guidelines mentioned in the checklist. A CONTRIBUTING.md file covering those points would do.

@prashjha, once that detail is fixed my review is done. @kokitsuyuzaki, thank you for developing a useful package!
