+++ CONTRIBUTING.md
@@ -1,0 +1,14 @@
+## Contributing
+Thanks for having the interest of this package.
+Your support is wellcome for keeping it great.
+
+Please note that this project is released with a [Contributor Code of Conduct][code-of-conduct].
+
+By participating in this project you agree to abide by its terms.
+
+## Issues
+Please follow [our guideline](.github/ISSUE_TEMPLATE/bug_report.md) to create an issue.
+
+## PRs
+For the details about how to create a pull request, see the official document of [GitHub](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request).
+
