+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,23 @@
+---
+name: Bug report
+about: Create a report to help us improve
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**To Reproduce**
+Please provide the minimum example to reproduce the error.
+
+**Session Information**
+Please paste the result of `sessionInfo()`.
+
+**Screenshots**
+If applicable, add screenshots or terminal output to help explain your problem.
+
+**Additional context**
+Add any other context about the problem here.
