### 2023-03-09 #5015 @kokitsuyuzaki
Thanks, @oneilsh.
I created my first CONTRIBUTING.md and learned a lot.
Please check the following markdowns.
https://github.com/rikenbit/nnTensor#contributing
https://github.com/rikenbit/nnTensor/blob/master/CONTRIBUTING.md
https://github.com/rikenbit/nnTensor/blob/master/.github/ISSUE_TEMPLATE/bug_report.md
https://github.com/rikenbit/nnTensor/blob/master/CODE_OF_CONDUCT.md
