### 2023-03-15 #5015 @kokitsuyuzaki
@Stat-Cook I have responded to your review.
Could you please check it?
https://github.com/openjournals/joss-reviews/issues/5015#issuecomment-1431484433
