### 2023-03-16 rikenbit/nnTensor#2 @Stat-Cook
Misleading behaviour in vignettes
In the vignettes when plotting the convergences they currently index as "[2:101]" - resulting in NA values at the end of the vector.

When I first checked this I was concerned there were numeric errors at the end of optimization, rather than the vectors aren't of this length.  I would suggest index instead as "[-1]" to drop the first (spurious) term and then the convergence vectors and plots are an accurate representation of the convergence.
