### 2023-03-16 #5015 @Stat-Cook
The recommendations from oneilsh covered some of my earlier concerns.  I'm happy the package does what it says on the tin - there's one recommendation I've added to the issues tracker for consideration but it doesn't effect functionality.  Happy to recommend acceptance.
