+++ vignettes/nnTensor-1.Rmd
@@ -59,8 +59,8 @@ The reconstruction error (`RecError`) and relative error (`RelChange`, the amoun
 
 ```{r conv_nmf, echo=TRUE, fig.height=4, fig.width=8}
 layout(t(1:2))
-plot(log10(out_NMF$RecError[2:101]), type="b", main="Reconstruction Error")
-plot(log10(out_NMF$RelChange[2:101]), type="b", main="Relative Change")
+plot(log10(out_NMF$RecError[-1]), type="b", main="Reconstruction Error")
+plot(log10(out_NMF$RelChange[-1]), type="b", main="Relative Change")
 ```
 
 The product of $U$ and $V$ shows that the original data can be well-recovered by `NMF`.
@@ -114,8 +114,8 @@ As same as NMF, the values of reconstruction error and relative error indicate t
 
 ```{r conv_nmtf, echo=TRUE, fig.height=4, fig.width=8}
 layout(t(1:2))
-plot(log10(out_NMTF$RecError[2:101]), type="b", main="Reconstruction Error")
-plot(log10(out_NMTF$RelChange[2:101]), type="b", main="Relative Change")
+plot(log10(out_NMTF$RecError[-1]), type="b", main="Reconstruction Error")
+plot(log10(out_NMTF$RelChange[-1]), type="b", main="Relative Change")
 ```
 
 The reconstructed matrix ($USV'$) shows that the features of the data matrix are well captured by NMTF.
