+++ vignettes/nnTensor-3.Rmd
@@ -63,8 +63,8 @@ As same as other matrix factorization methods, the values of reconstruction erro
 
 ```{r conv_ntf_cp, echo=TRUE, fig.height=4, fig.width=8}
 layout(t(1:2))
-plot(log10(out_NTF_CP$RecError[2:101]), type="b", main="Reconstruction Error")
-plot(log10(out_NTF_CP$RelChange[2:101]), type="b", main="Relative Change")
+plot(log10(out_NTF_CP$RecError[-1]), type="b", main="Reconstruction Error")
+plot(log10(out_NTF_CP$RelChange[-1]), type="b", main="Relative Change")
 ```
 
 By using `recTensor`, user can easily reconstruct the data from core tensor and factor matrices as follows.
@@ -86,8 +86,8 @@ str(out_NTF_Tucker, 2)
 
 ```{r conv_ntf_tucker, echo=TRUE, fig.height=4, fig.width=8}
 layout(t(1:2))
-plot(log10(out_NTF_Tucker$RecError[2:101]), type="b", main="Reconstruction Error")
-plot(log10(out_NTF_Tucker$RelChange[2:101]), type="b", main="Relative Change")
+plot(log10(out_NTF_Tucker$RecError[-1]), type="b", main="Reconstruction Error")
+plot(log10(out_NTF_Tucker$RelChange[-1]), type="b", main="Relative Change")
 ```
 
 Unlike with the first data, the second data shows that `NTF` does not work.
@@ -111,8 +111,8 @@ str(out_NTD, 2)
 
 ```{r conv_ntd, echo=TRUE, fig.height=4, fig.width=8}
 layout(t(1:2))
-plot(out_NTD$RecError[2:101], type="b", main="Reconstruction Error")
-plot(out_NTD$RelChange[2:101], type="b", main="Relative Change")
+plot(out_NTD$RecError[-1], type="b", main="Reconstruction Error")
+plot(out_NTD$RelChange[-1], type="b", main="Relative Change")
 ```
 
 ```{r rec_ntd_tucker, echo=TRUE, fig.height=4, fig.width=8}
@@ -135,12 +135,12 @@ out_NTD2_3 <- NTD(X_Tucker, rank=c(4,5), modes=2:3)
 
 ```{r conv_ntd2, echo=TRUE, fig.height=12, fig.width=8}
 layout(rbind(1:2, 3:4, 5:6))
-plot(out_NTD2_1$RecError[2:101], type="b", main="Reconstruction Error\nNTD-2 (mode=1:2)")
-plot(out_NTD2_1$RelChange[2:101], type="b", main="Relative Change\nNTD-2 (mode=1:2)")
-plot(out_NTD2_2$RecError[2:101], type="b", main="Reconstruction Error\nNTD-2 (mode=c(1,3))")
-plot(out_NTD2_2$RelChange[2:101], type="b", main="Relative Change\nNTD-2 (mode=c(1,3))")
-plot(out_NTD2_3$RecError[2:101], type="b", main="Reconstruction Error\nNTD-2 (mode=2:3)")
-plot(out_NTD2_3$RelChange[2:101], type="b", main="Relative Change\nNTD-2 (mode=2:3)")
+plot(out_NTD2_1$RecError[-1], type="b", main="Reconstruction Error\nNTD-2 (mode=1:2)")
+plot(out_NTD2_1$RelChange[-1], type="b", main="Relative Change\nNTD-2 (mode=1:2)")
[…]
