### 2023-03-17 rikenbit/nnTensor#2 @kokitsuyuzaki
Thanks, I replaced all the [2:101] to [-1] in v1.1.13.
https://github.com/rikenbit/nnTensor/commit/1257383c32035c31951e44c7dd98c7dd8052cc32
