### 2023-03-17 #5015 @kokitsuyuzaki
@Stat-Cook Thank you for checking the details. The vignette has been fixed in v1.1.13.
https://github.com/rikenbit/nnTensor/issues/2
