### 2023-04-17 #5015 @prashjha
Hi @kokitsuyuzaki, your draft looks good. Could you archive (if not done already) archive the release using zenodo and provide the archive reference so that I can associate it with your JOSS submission? Make sure that the zenodo archive's title matches this JOSS submission's title.

Once you are done, I will hand your paper to EiC for the final decision.
