### 2023-04-18 #5015 @kokitsuyuzaki
Thanks, @prashjha
I achieved the release and the Zenodo DOI is 10.5281/zenodo.7839285
Please check the links below.

https://zenodo.org/record/7839285#.ZD32Hy33JhE
https://github.com/rikenbit/nnTensor/releases/tag/joss_released
