### 2023-04-18 #5015 @prashjha
Hi @kokitsuyuzaki, could you please attach the latest version (formatted like vx.x.x or x.x.x) to the Zenodo archive? I see the latest version attached to your zenoda archive as 'joss_released'. This is a nonstandard version name. Once  you update zenodo, let me know the archive DOI and also the corresponding released version of your repository so that I can associate these details with your JOSS article. Thanks.
