### 2023-04-20 #5015 @kokitsuyuzaki
Sorry, I've created new tag v1.1.14 and the Zenodo DOI is 10.5281/zenodo.7848088.

https://zenodo.org/record/7848088#.ZEDeni33JhE
https://github.com/rikenbit/nnTensor/releases/tag/v1.1.14
